Create the reset button at the top of the grid
When click on the reset button, the user will be prompted to enter
a new number of squares per side. After having the new number of
squares, the current grid will be removed and a new grid will be
created based on the new number of squares entered by the user.